pyproject.toml

[project]
name = "Inforratica-Django"
version = "0.1.0"
description = ""
authors = [
    {name = "Guilherme Tamanini", email = "guilherme.gui.tama@gmail.com"},
]
dependencies = [
    "django>=4.2.4",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.0.0",
    "setuptools>=68.0.0",
    "django-cpf>=0.1.0",
    "djangorestframework-simplejwt>=5.2.2",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.3.0",
]
